backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.7
python-dotenv>=1.0.0
opencv-python==4.9.0.80
numpy==1.26.3
torch==2.5.1
torchvision==0.20.1
ultralytics==8.1.0
supabase>=2.3.0
